Add files via upload
Partner, I have made some changes and optimizations to your code, this is the optimized version, also I was wondering if according to the scoring criteria we only need to implement the wavefront propagation algorithm to be able to get the score. I have tried many modifications and it seems that it is difficult to solve difficult sudoku quickly using only the wavefront propagation algorithm, and sometimes it even fails to do so. Perhaps it is enough that we only need to implement the wavefront propagation algorithm?